requirements.txt
Django==1.11.4
psycopg2==2.7.3
whitenoise==3.3.0
djangorestframework==3.6.3
django-rest-swagger==2.1.2
waitress==1.0.2
django-oauth-toolkit==1.0.0
django-cors-middleware==1.3.1
django-elasticsearch-dsl==0.2.0
elasticsearch-dsl==5.3.0
dj-database-url==0.4.2
social-auth-core==1.4.0
social-auth-app-django==1.2.0
django-extensions==1.9.0
